+++ ServidorV2/.settings/org.eclipse.wst.common.project.facet.core.prefs.xml
@@ -1,0 +1,7 @@
+<root>
+  <facet id="jpt.jpa">
+    <node name="libprov">
+      <attribute name="provider-id" value="jpa-no-op-library-provider"/>
+    </node>
+  </facet>
+</root>
