+++ ServidorV2/.settings/org.eclipse.wst.common.project.facet.core.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<faceted-project>
+  <installed facet="java" version="1.8"/>
+  <installed facet="jpt.jpa" version="2.0"/>
+</faceted-project>
