+++ ServidorV2/src/main/java/com/DB/ChuckNorris.java
@@ -2,18 +2,27 @@ package com.DB;
 
 import java.io.Serializable;
 
+import javax.persistence.Column;
 import javax.persistence.Entity;
+import javax.persistence.Id;
 import javax.persistence.Table;
 
 @Entity
-@Table(name = "ROBOTICA")
-public class Robotica implements Serializable {
+@Table(name = "CHUCK")
+public class ChuckNorris implements Serializable {
 
-	private static final long serialVersionUID = -371128921165058773L;
+	private static final long serialVersionUID = -5153794647898719855L;
+	
+	@Id
+	@Column(name="Chuck_ID")
 	private int ID;
+	
+	@Column(name="FACT")
 	private String frase;
+	
+	public ChuckNorris() {}
 
-	public Robotica(int iD, String frase) {
+	public ChuckNorris(int iD, String frase) {
 		super();
 		ID = iD;
 		this.frase = frase;
@@ -34,5 +43,4 @@ public class Robotica implements Serializable {
 	public void setFrase(String frase) {
 		this.frase = frase;
 	}
-
 }
